run: Implement fuzzy ref matching
We've had the feature for a long time that allows you to use a substring
instead of a full app ID in the install and uninstall commands. The only
reason I didn't add it to the run command back then was that I wasn't
sure it was ok to add interactive prompts to flatpak-run (it is often
invoked by other programs rather than users directly). But in fact we
already have a prompt in case there's ambiguity about which branch to
use; this commit just repurposes that to also address ambiguity about
which ref to use.

Long story short, this allows you to do "flatpak run firefox" instead of
"flatpak run org.mozilla.firefox".

Helps: #1258